+++ src/main/java/jamdoggie/betterbattletowers/worldgen/data/adapter/BlockPalletJsonAdapter.java
@@ -1,0 +1,4 @@
+package jamdoggie.betterbattletowers.worldgen.data.adapter;
+
+public class BlockPalletJsonAdapter {
+}
