+++ src/main/java/jamdoggie/betterbattletowers/worldgen/data/decoration/BlockData.java
@@ -1,4 +1,4 @@
-package jamdoggie.betterbattletowers.worldgen.properties.decoration;
+package jamdoggie.betterbattletowers.worldgen.data.decoration;
 
 import java.util.Objects;
