--- src/main/java/jamdoggie/betterbattletowers/worldgen/properties/adapter/BlockPalletJsonAdapter.java
@@ -1,4 +1,0 @@
-package jamdoggie.betterbattletowers.worldgen.properties.adapter;
-
-public class BlockPalletJsonAdapter {
-}
